// src/types/enums/Action.ts
export enum Action {
  TOGGLE_SHOWMODAL = 'TOGGLE_SHOWMODAL',
  TOGGLE_LOWERCASE = 'TOGGLE_LOWERCASE',
  TOGGLE_UPPERCASE = 'TOGGLE_UPPERCASE',
  TOGGLE_NUMBERS = 'TOGGLE_NUMBERS',
  TOGGLE_SYMBOLES = 'TOGGLE_SYMBOLES',
  TOGGLE_EASY = 'TOGGLE_EASY',
  TOGGLE_MEDIUM = 'TOGGLE_MEDIUM',
  TOGGLE_HARD = 'TOGGLE_HARD',
  TOGGLE_CONTROLED = 'TOGGLE_CONTROLED',
  ADD_NOTE = 'ADD_NOTE',
  DELETE_NOTE = 'DELETE_NOTE',
  SHOW_PASSWORD = 'SHOW_PASSWORD',
  TOGGLE_LENGTH = 'TOGGLE_LENGTH',
  UPDATE_PASSWORD = 'UPDATE_PASSWORD',
  COPY_PASSWORD = 'COPY_PASSWORD'
}